first and the last comment, nodejs basic server format
index.js is like the setting before we start server.
server.js is like the config.conf in Ruby on Rails.
router.js is like the router.rb in Ruby on Rails.
controller.js, actually it named "requestHandler.js" in guide, but I think
it's pretty similar to the controller.rb, so I name it like that.

use fs in controller.js to show .html file